Cleave variogram functions from `spatial.py` and move into `variogram.py`
