+++ echopop/nwfsc_feat/spatial.py
@@ -1,10 +1,9 @@
 import warnings
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, Optional, Tuple, Union
 
 import geopandas as gpd
 import numpy as np
 import pandas as pd
-from lmfit import Minimizer, Parameters
 from scipy import interpolate
 from shapely.geometry import Point, Polygon
 from shapely.ops import unary_union
@@ -91,566 +90,6 @@ def transform_coordinates(
 
     # Return the output tuple
     return (data, delta_x, delta_y)
-
-
-def lag_distance_matrix(
-    coordinates_1: Union[pd.DataFrame, np.ndarray],
-    coordinate_names: Optional[Tuple[str, str]] = None,
-    coordinates_2: Optional[Union[pd.DataFrame, np.ndarray]] = None,
-    self: bool = False,
-    azimuth_matrix: bool = False,
-) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Calculate the lag distance matrix between two sets of coordinates.
-
-    Parameters
-    ----------
-    coordinates_1 : pd.DataFrame or np.ndarray
-        First set of coordinates (x, y) as a DataFrame or numpy array.
-    coordinate_names : Tuple[str, str], optional
-        Names of the coordinate columns when using DataFrames. Expected format: (x_col, y_col).
-    coordinates_2 : pd.DataFrame or np.ndarray, optional
-        Second set of coordinates (x, y) as a DataFrame or numpy array. If None, uses coordinates_1.
-    self : bool, default=False
-        If True, calculates distances within the same set of coordinates.
-    azimuth_matrix : bool, default=False
-        If True, returns both distance and azimuth angles; if False, returns only distances.
-
-    Returns
-    -------
-    Tuple[np.ndarray, np.ndarray]
-        If azimuth_matrix is True, returns a tuple of (distance matrix, azimuth angles).
-        If azimuth_matrix is False, returns (distance matrix, empty array).
-    """
-
-    # Get coordinate names
-    if coordinate_names is not None:
-        x_name, y_name = coordinate_names
-
-    # Set reference to self if 'coordinates_2' is not defined
-    if coordinates_2 is None:
-        coordinates_2 = coordinates_1
-
-    # Get the distance array coordinates in the 'x' and 'y' directions
-    # ---- Case: Internal distances (as arrays) and self is True
-    if self and all(isinstance(x, np.ndarray) for x in [coordinates_1, coordinates_2]):
-        # ---- x-coordinates
-        x_coords = (coordinates_1, coordinates_1)
-        # ---- y-coordinates
-        y_coords = (coordinates_2, coordinates_2)
-    # ---- Case: Distances (as arrays) and self is False
-    elif not self and all(isinstance(x, np.ndarray) for x in [coordinates_1, coordinates_2]):
-        # ---- x-coordinates
-        x_coords = (coordinates_1, coordinates_2)
-        # ---- y-coordinates
-        y_coords = (coordinates_1, coordinates_2)
-    # ---- Case: DataFrames
-    elif all(isinstance(x, pd.DataFrame) for x in [coordinates_1, coordinates_2]):
-        # ---- x-coordinates
-        x_coords = (coordinates_1[x_name].to_numpy(), coordinates_2[x_name].to_numpy())
-        # ---- y-coordinates
-        y_coords = (coordinates_1[y_name].to_numpy(), coordinates_2[y_name].to_numpy())
-    #  Resolve the distances
-    # ---- x-distance
-    x_distance = np.subtract.outer(*x_coords)
-    # ---- y-distance
-    y_distance = np.subtract.outer(*y_coords)
-
-    # Get the azimuth angle matrix, if required
-    if azimuth_matrix:
-        # ---- Create copies of 'x_distance' and 'y_distance'
-        x_angles = x_distance.copy()
-        y_angles = y_distance.copy()
-        # ---- Replace the self-points with 'NaN'
-        np.fill_diagonal(x_angles, np.nan)
-        np.fill_diagonal(y_angles, np.nan)
-        # ---- Calculate the azimuth angle grid
-        azimuth_grid = (
-            np.arctan(
-                np.divide(y_angles, x_angles, where=(x_angles != 0.0) & (~np.isnan(x_angles)))
-            )
-            * 180.0
-            / np.pi
-            + 180.0 % 180.0
-        )
-        # ---- Return the resulting tuple of matrices for Euclidean distances and azimuth angles
-        return np.sqrt(x_distance * x_distance + y_distance * y_distance), azimuth_grid
-    else:
-        # ---- Return Euclidean distance matrix
-        return np.sqrt(x_distance * x_distance + y_distance * y_distance), np.array([])
-
-
-def filter_lag_matrix(
-    data_matrix: np.ndarray[int],
-    mask_matrix: np.ndarray[bool],
-    azimuth_matrix: Optional[np.ndarray[float]] = None,
-    azimuth_angle_threshold: Optional[float] = None,
-) -> np.ndarray[int]:
-    """
-    Filter the lag matrix based on a boolean mask and optional azimuth angle matrix.
-
-    Parameters
-    ----------
-    data_matrix : np.ndarray
-        The lag distance matrix to be filtered.
-    mask_matrix : np.ndarray[bool]
-        A boolean mask indicating which elements to keep in the lag matrix. This is typically a
-        triangle boolean matrix.
-    azimuth_matrix : np.ndarray[float], optional
-        An optional azimuth angle matrix that can be used to filter the lag matrix based on
-        azimuth angles.
-    azimuth_angle_threshold : float, optional
-        If provided, this threshold is used to filter the azimuth angles in the azimuth matrix.
-        This defines the total azimuth angle range that is allowed for constraining the
-        relative angles between spatial points, particularly for cases where a high degree of
-        directionality is assumed.
-
-    Returns
-    -------
-    np.ndarray
-        The filtered lag matrix, where elements not meeting the mask or azimuth criteria are
-        extracted as a 1D array.
-    """
-
-    # Convert array to matrix, if needed
-    if data_matrix.ndim == 1:
-        data_matrix = np.tile(data_matrix, (len(data_matrix), 1))
-    else:
-        if data_matrix.shape != mask_matrix.shape:
-            # ---- Determine which dimension is mismatched
-            dimension_diff = np.where(np.array(data_matrix.shape) != np.array(mask_matrix.shape))[0]
-            if dimension_diff == 0:
-                data_matrix = np.tile(data_matrix, (len(data_matrix), 1))
-            else:
-                data_matrix = np.tile(data_matrix, (1, len(data_matrix)))
-
-    # If 'azimuth_matrix' is supplied, then apply threshold as additional bitmap
-    if (
-        azimuth_matrix is not None
-        and len(azimuth_matrix) > 0
-        and azimuth_angle_threshold is not None
-    ):
-        # ---- Replace any 'NaN' values with 0's
-        azimuth_matrix[np.isnan(azimuth_matrix)] = 0.0
-        # ---- Create bitmap
-        azimuth_mask = (azimuth_matrix >= -azimuth_angle_threshold) & (
-            azimuth_matrix < azimuth_angle_threshold
-        )
-    else:
-        # ---- Create empty azimuth mask
-        azimuth_mask = np.ones_like(data_matrix, dtype=bool)
-
-    # Mask the data matrix and broadcast into a 1D array
-    return data_matrix[mask_matrix & azimuth_mask]
-
-
-def quantize_lags(
-    estimates: np.ndarray[float],
-    lag_matrix: np.ndarray[int],
-    mask_matrix: np.ndarray[bool],
-    azimuth_matrix: np.ndarray[float],
-    n_lags: int,
-    azimuth_angle_threshold: Optional[float] = None,
-) -> Tuple[np.ndarray[int], np.ndarray[float], np.ndarray[float], np.ndarray[float]]:
-    """
-    Quantize lag counts, sums, square-sums, and deviations required for computing the empirical
-    variogram
-
-    Parameters
-    ----------
-    estimates : np.ndarray[float]
-        A 1D array of field estimates, e.g. 'biomass'.
-    lag_matrix : np.ndarray[int]
-        A 2D array of integer lag values.
-    mask_matrix : np.ndarray[bool]
-        A boolean mask indicating which elements to keep in the lag matrix. This is typically a
-        triangle boolean matrix.
-    azimuth_matrix : np.ndarray[float]
-        A 2D array of azimuth angle values. Alternatively, this may also be a 1D array of length 0
-        for the case where the azimuth angle matrix is not defined.
-    n_lags : int
-        Number of lags used for the variogram analysis.
-    azimuth_angle_threshold : float, optional
-        If provided, this threshold is used to filter the azimuth angles in the azimuth matrix.
-        This defines the total azimuth angle range that is allowed for constraining the
-        relative angles between spatial points, particularly for cases where a high degree of
-        directionality is assumed.
-
-    Returns
-    -------
-    Tuple[np.ndarray[int], np.ndarray[float], np.ndarray[float], np.ndarray[float]]
-        A tuple comprising the binned lag counts, summed estimates, square-summed estimates, and
-        deviations.
-    """
-
-    # Validate that `estimates` is a 1D array
-    if estimates.ndim > 1:
-        raise ValueError("Estimates array ('estimates') must be a 1D array.")
-
-    # Validate that dimensions are 2D
-    if (
-        lag_matrix.ndim < 2
-        or mask_matrix.ndim < 2
-        or (azimuth_matrix.ndim < 2 and len(azimuth_matrix) > 0)
-    ):
-        error = (
-            "The function `quantize_lags` requires arrays to be 2D. The following 1D arrays have "
-            "invalid shapes: "
-        )
-        invalid_arrays = []
-        if lag_matrix.ndim < 2:
-            invalid_arrays += ["'lag_matrix'"]
-        if mask_matrix.ndim < 2:
-            invalid_arrays += ["'mask_matrix'"]
-        if azimuth_matrix.ndim < 2 and len(azimuth_matrix) > 0:
-            invalid_arrays += ["'azimuth_matrix'"]
-        raise ValueError(error + ", ".join(invalid_arrays))
-
-    # Filter the lag matrix based on the mask and azimuth angle threshold
-    equivalent_lags = filter_lag_matrix(
-        lag_matrix, mask_matrix, azimuth_matrix, azimuth_angle_threshold
-    )
-    # ---- Compute the binned sums
-    lag_counts = np.bincount(equivalent_lags)[1:n_lags]
-
-    # Sum the estimates for each lag
-    estimates_lagged = filter_lag_matrix(
-        estimates, mask_matrix, azimuth_matrix, azimuth_angle_threshold
-    )
-    # ---- Compute the binned sums
-    lag_estimates = np.bincount(equivalent_lags, weights=estimates_lagged)[1:n_lags]
-
-    # Compute the binned squared-sum
-    lag_estimates_squared = np.bincount(equivalent_lags, weights=estimates_lagged**2)[1:n_lags]
-
-    # Generate a lag bitmap
-    lag_bitmap = equivalent_lags < n_lags
-
-    # Filter the estimates matrix
-    estimates_matrix = filter_lag_matrix(
-        np.arange(len(estimates))[:, np.newaxis],
-        mask_matrix,
-        azimuth_matrix,
-        azimuth_angle_threshold,
-    )
-
-    # Calculate the deviations between the indexed and lag-specific estimates
-    deviations = (estimates[estimates_matrix][lag_bitmap] - estimates_lagged[lag_bitmap]) ** 2
-
-    # Sum the deviations per lag bin
-    lag_deviations = np.bincount(equivalent_lags[lag_bitmap], weights=deviations)[1:n_lags]
-
-    # Return the calculated quantities
-    return lag_counts, lag_estimates, lag_estimates_squared, lag_deviations
-
-
-def semivariance(
-    estimates: np.ndarray[float],
-    lag_estimates: np.ndarray[float],
-    lag_estimates_squared: np.ndarray[float],
-    lag_counts: np.ndarray[int],
-    lag_deviations: np.ndarray[float],
-    head_index: np.ndarray[int],
-) -> Tuple[np.ndarray[float], np.ndarray[float]]:
-    """
-    Compute the (standardized) semivariance
-
-    Parameters
-    ----------
-    estimates : np.ndarray[float]
-        A 1D array of field estimates, e.g. 'biomass'.
-    lag_estimates : np.ndarray[float]
-        Summed lag estimates weighted by values for the values within 'estimates'.
-    lag_estimates_squared : np.ndarray[float]
-        Square-summed lag estimates weighted by values for the values within 'estimates'.
-    lag_counts : np.ndarray[int]
-        Binned counts of values within each lag.
-    lag_deviations : np.ndarray[float]
-        Statistical deviation within each lag bin.
-    head_index : np.ndarray[int]
-        A 2D array containing the head indices of each lag for each row.
-
-    Returns
-    -------
-    Tuple[np.ndarray[float], np.ndarray[float]]
-        A tuple comprising 1D arrays with the semivariance estimates and mean lag covariance at
-        each lag.
-    """
-
-    # Calculate the mean head estimate per lag bin
-    mean_head = (estimates[:, np.newaxis] * (head_index / lag_counts)).sum(axis=0)
-
-    # Calculate the standard deviation of head values per lag
-    sigma_head = np.sqrt(
-        ((estimates[:, np.newaxis] - mean_head) ** 2 * (head_index / lag_counts)).sum(axis=0)
-    )
-
-    # Calculate the global mean and variance for each lag bin
-    # ---- Mean
-    lag_means = lag_estimates / lag_counts
-    # ---- Variance
-    lag_variance = lag_estimates_squared / lag_counts - lag_means**2
-
-    # Estimate the standard deviation of tail estimates
-    sigma_tail = np.sqrt(np.abs(lag_variance))
-
-    # Calculate the semivariance
-    # ---- Compute the partial sill that is applied as a weighted calculation
-    partial_sill = sigma_tail * sigma_head
-    # ---- Semivariance [gamma(h)] and cross-sill estimate
-    # gamma_h = 0.5 * lag_deviations / (lag_counts * partial_sill)
-    with np.errstate(divide="ignore"):
-        gamma_h = 0.5 * lag_deviations / (lag_counts * partial_sill)
-    # Calculate the mean lag distance covariance
-    # ---- Find non-zero head and tail variances
-    non_zero_variance = np.where((sigma_head > 0.0) & (sigma_tail > 0.0))[0]
-    # ---- Mean lag covariance
-    if non_zero_variance.size > 0:
-        mean_lag_covariance = (sigma_head[non_zero_variance] * sigma_tail[non_zero_variance]).mean()
-    else:
-        mean_lag_covariance = np.nan
-
-    # Return the outputs
-    return gamma_h, mean_lag_covariance
-
-
-def empirical_variogram(
-    data: pd.DataFrame,
-    n_lags: int,
-    lag_resolution: float,
-    azimuth_filter: bool,
-    azimuth_angle_threshold: float,
-    variable: str = "biomass_density",
-    coordinate_names: Tuple[str, str] = ("x", "y"),
-    force_lag_zero: bool = True,
-) -> Tuple[np.ndarray[float], np.ndarray[float], np.ndarray[int], float]:
-    """
-    Compute the empirical variogram from transect data.
-
-    Parameters
-    ----------
-    data : pd.DataFrame
-        A dataframe containing georeferenced coordinates associated with a particular variable (e.g.
-        biomass). This DataFrame must have at least two valid columns comprising the overall 2D
-        coordinates (e.g. 'x' and 'y').
-    n_lags : int
-        The number of lags used for computing the (semi)variogram.
-    lag_resolution : float
-        The distance interval represented by each lag interval.
-    azimuth_filter : bool
-        When True, a 2D array of azimuth angles are generated. This subsequent array represents the
-        relative azimuth angles between spatial points, and can serve as a filter for case where
-        a high degree of directionality is assumed. This accompanies the argument
-        'azimuth_angle_threshold' that defines the threshold azimuth angle.
-    azimuth_angle_threshold : float
-        This threshold is used for filtering the azimuth angles.
-    variable : str, default = 'biomass_density'
-        The variable used for computing the empirical variogram (e.g. 'biomass_density'), which
-        must exist as a column in 'transect_df'.
-    coordinate_names : Tuple[str, str], default = ('x', 'y')
-        A tuple containing the 'transect_df' column names defining the coordinates. The order of
-        this input matters where they should be defined as the (horizontal axis, vertical axis).
-    force_lag_zero : bool, default = True
-        When True, the nugget effect is assumed to be 0.0 for the empirical variogram. This adds
-        lag 0 to the subsequent array outputs where semivariance (or 'gamma_h') is also equal to
-        0.
-
-    Returns
-    -------
-    Tuple[np.ndarray[float], np.ndarray[float], np.ndarray[int], float]
-        A tuple containing arrays with the lag intervals, semivariance, and lag counts. The mean
-        lag covariance between the head and tail points computed for all input data is also
-        provided.
-
-    """
-    # Initialize lag array
-    lags = np.concatenate([np.arange(1, n_lags) * lag_resolution])
-
-    # Calculate the distance (and azimuth) matrix
-    distance_matrix, azimuth_matrix = lag_distance_matrix(
-        coordinates_1=data,
-        coordinate_names=coordinate_names,
-        self=True,
-        azimuth_matrix=azimuth_filter,
-    )
-
-    # Convert lag distances to lag numbers
-    lag_matrix = np.round(distance_matrix / lag_resolution).astype(int) + 1
-
-    # Extract estimates column
-    estimates = data[variable].to_numpy()
-
-    # Create a triangle mask with the diaganol offset to the left by 1
-    # ---- Initial mask
-    triangle_mask = np.tri(len(estimates), k=-1, dtype=bool)
-    # ---- Vertically and then horizontally flip to force the 'True' and 'False' positions
-    triangle_mask_flp = np.flip(np.flip(triangle_mask), axis=1)
-
-    # Quantize the lags
-    lag_counts, lag_estimates, lag_estimates_squared, lag_deviations = quantize_lags(
-        estimates, lag_matrix, triangle_mask_flp, azimuth_matrix, n_lags, azimuth_angle_threshold
-    )
-
-    # Compute the mean and standard deviation of the head estimates for each lag bin
-    # ---- Apply a mask using the triangle bitmap
-    head_mask = np.where(triangle_mask_flp, lag_matrix, -1)
-
-    # Helper function for computing the binned summations for each row
-    def bincount_row(row, n_lags):
-        return np.bincount(row[row != -1], minlength=n_lags)[1:n_lags]
-
-    # Pre-allocate vectors/arrays that will be iteratively filled
-    head_index = np.zeros((len(estimates), n_lags - 1))
-    # ---- Find the head indices of each lag for each row
-    head_index = np.apply_along_axis(bincount_row, axis=1, arr=head_mask, n_lags=n_lags)
-
-    # Compute the standardized semivariance [gamma(h)]
-    gamma_h, lag_covariance = semivariance(
-        estimates, lag_estimates, lag_estimates_squared, lag_counts, lag_deviations, head_index
-    )
-
-    # Prepend a 0.0 and force the nugget effect to be 0.0, if necessary
-    # ---- Return the computed lags, empirical variogram estimate [gamma(h)], and lag counts
-    if force_lag_zero:
-        return (
-            np.concatenate([[0], lags]),
-            np.concatenate([[0.0], gamma_h]),
-            np.concatenate([[len(estimates) - 1], lag_counts]),
-            lag_covariance,
-        )
-    else:
-        return lags, gamma_h, lag_counts, lag_covariance
-
-
-def fit_variogram(
-    lags: np.ndarray[float],
-    lag_counts: np.ndarray[int],
-    gamma: np.ndarray[float],
-    variogram_parameters: Parameters,
-    model: Union[str, List[str]] = ["bessel", "exponential"],
-    optimizer_kwargs: Dict[str, Any] = {},
-) -> Tuple[Dict[str, Any], float, float]:
-    """
-    Compute the best-fit variogram parameters for input data
-
-    Parameters
-    ----------
-    lags : np.ndarray[float]
-        A 1D array of the lag distances.
-    lag_counts : np.ndarray[int]
-        A 1D array of the lag counts.
-    gamma : np.ndarray[float]
-        A 1D array comprising the semivariance estimates for each lag.
-    variogram_parameters : Parameters
-        A :class:`lmfit.parameter.Parameters` object containing the parameters required for the
-        defined variogram model. See :func:`echopop.spatial.variogram.variogram` for more details.
-        Valid parameters include:
-
-            - **sill : float**
-            The asymptotic value as lags approach infinity
-
-            - **nugget : float**
-            The semivariogram *y*-intercept that corresponds to variability at lag distances
-            shorter than the lag resolution
-
-            - **correlation_range : float**
-            The relative length scale, or range, at which the autocorrelation between lag distances
-            no longer increases and becomes asymptotic
-
-            - **hole_effect_range : float**
-            The (normalized) length scale/range that holes' are observed, which represent 'null'
-            (or very small) points compared to their neighboring lags
-
-            - **decay_power : float**
-            An exponential term that is used in certain generalized exponential (or related)
-            semivariogram models that modulates the ascending rate for a semivariogram
-
-            - **enhance_semivariance : bool**
-            A boolean term that determines whether the correlation decay in certain  cosine-related
-            variogram models are enhanced (or not) with increasing lag distances
-
-    model : Union[str, List[str]], default=['bessel', 'exponential']
-        A string or list of model names. A single name represents a single family model. Two inputs
-        represent the desired composite model (e.g. the composite J-Bessel and exponential model).
-        Defaults to: ``model=["bessel", "exponential"]``. Available models and their required
-        arguments can be reviewed in the :func:`echopop.spatial.variogram.variogram` function.
-    optimizer_kwargs : Dict[str, Any], default={}
-        A dictionary comprising the various function arguments used by
-        :class:`lmfit.minimizer.Minimizer` for Least-Squares minimization that incorporates the
-        Trust Reflective method. See :class:`lmfit.minimizer.Minimizer` for more details.
-
-    Returns
-    -------
-    Tuple[Dict[str, Any], float, float]
-        A tuple containing a dictionary with the best-fit keyword variogram parameter values,
-        the mean absolute deviation (MAD) of the initial parameter values, and the MAD of the
-        best-fit parameter values.
-
-    See Also
-    --------
-    :func:`echopop.spatial.variogram.variogram` :
-        Variogram model parameters.
-    :class:`lmfit.parameter.Parameters` :
-        Variogram parameter optimization leverages the ``Parameters`` class from ``lmfit`` for
-        model optimization.
-    :class:`lmfit.minimizer.Minimizer` :
-        Optimization keyword arguments used for the Least-Squares fitting.
-    """
-    # Normalize the lag counts to get the lag weights
-    lag_weights = lag_counts / lag_counts.sum()
-
-    # Vertically stack the lags, semivariance, and weights
-    data_stack = np.vstack((lags, gamma, lag_weights))
-
-    # Recover the lag resolution
-    delta_lag = np.diff(lags).mean()
-
-    # Compute the range
-    range = lags.max() + delta_lag
-
-    # Index lag distances that are within the parameterized range
-    within_range = np.where(lags <= range)[0]
-
-    # Truncate the data stack
-    truncated_stack = data_stack[:, within_range]
-
-    # Create helper cost-function that is weighted using the kriging weights (`w`), lag
-    # distances (`x`), and empirical semivariance (`y`)
-    def cost_function(parameters, x, y, w, model):
-        yr = variogram(x, {**parameters, **{"model": model}})
-        return (yr - y) * w
-
-    # Compute the initial fit based on the pre-optimized parameter values
-    initial_fit = cost_function(
-        variogram_parameters,
-        x=truncated_stack[0],
-        y=truncated_stack[1],
-        w=truncated_stack[2],
-        model=model,
-    )
-
-    # Compute the initial mean absolute deviation (MAD)
-    mad_initial = np.mean(np.abs(initial_fit))
-
-    # Generate `Minimizer` function class required for bounded optimization
-    minimizer = Minimizer(
-        cost_function,
-        variogram_parameters,
-        fcn_args=(truncated_stack[0], truncated_stack[1], truncated_stack[2], model),
-    )
-
-    # Minimize the cost-function to compute the best-fit/optimized variogram parameters
-    parameters_optimized = minimizer.minimize(method="least_squares", **optimizer_kwargs)
-
-    # Calculate the optimized MAD
-    mad_optimized = np.mean(np.abs(parameters_optimized.residual))
-
-    # Extract the best-fit parameter values
-    best_fit_params = parameters_optimized.params.valuesdict()
-
-    # Return the final tuple
-    return best_fit_params, mad_initial, mad_optimized
-
 
 def get_survey_western_extents(
     transect_df: pd.DataFrame,
